user model, givelocationaccess, user route bug fix
user model change phone datatype, givelocationaccess send apt response back if not authorized.

+++ models/user.sql
@@ -12,7 +12,7 @@ create table users(
   email varchar(60) not null UNIQUE,
   password varchar(300) not null,
   name varchar(50) not null,
-  phone int not null,
+  phone bigint not null,
 );
 GO
